requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
Automat==24.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.37.4
botocore==1.37.4
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudscraper==1.2.71
constantly==23.10.4
cryptography==44.0.1
cssselect==1.2.0
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.17.0
frozenlist==1.5.0
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.28.0
numpy==2.2.3
openai==0.28.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.241126
parsel==1.10.0
pillow==11.1.0
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
propcache==0.2.1
Protego==0.4.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
queuelib==1.7.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
s3transfer==0.11.3
Scrapy==2.12.0
selenium==4.28.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
streamlit==1.42.2
tenacity==9.0.0
tldextract==5.1.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.0
Twisted==24.11.0
types-pytz==2025.1.0.20250204
typing_extensions==4.12.2
tzdata==2025.1
undetected-chromedriver==3.5.5
urllib3==2.3.0
w3lib==2.3.1
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
